создать презентер для film-details
