+++ src/model/films-model.js
@@ -48,22 +48,4 @@ export default class FilmsModel extends Observable {
 
     this._notify(updateType, update);
   };
-
-  addFilm = (updateType, update) => {
-    this.#films = [update, ...this.#films];
-
-    this._notify(updateType, update);
-  };
-
-  deleteFilm = (updateType, update) => {
-    const index = this.#films.findIndex((film) => film.id === update.id);
-
-    if (index === -1) {
-      throw new Error('Can\'t delete unexisting film');
-    }
-
-    this.#films = [...this.#films.slice(0, index), ...this.#films.slice(index + 1)];
-
-    this._notify(updateType);
-  };
 }
